feat: modularize component documentation structure and enhance content organization
